requirements.txt
Django==4.2.24
djangorestframework==3.16.0
django-cors-headers==4.8.0
Pillow==10.2.0
